use header and source for bip38
